refactor: Azure SD, remove SD struct from config
